iOS: Move swizzles out of FullStory.framework, and implement fsTagName and fsAttribute. (#11)
Previously, FullStory.framework was responsible for swizzling RCTViewManager to add new properties if we were running a React app.  Now that there is a native element to the FullStory React Native plugin, instead of swizzling, we can provide those as an `@implementation` of RCTViewManager.

Additionally, as long as we're moving these around, add support for the new React view attributes.