Resolve UniModules incompatibility by swizzling [RCTComponentData viewConfig]. (#15)
UniModules gets confused about FullStory properties; it generates a list of
'proxiedProperties' that, in theory, excludes properties that come from React
Native.  Unfortunately, it doesn't see the FullStory properties, and takes
those out of the main tag, and passes them in the proxiedProperties struct to
the UniModule view.  This means that we don't get to see the property -- and
that UniModule generates an irritated warning about this.

We solve this by swizzling [RCTComponentData viewConfig] to inject the
FullStory properties into each view's list of NativeProps.  We need to do this
because NativeProps uses class_copyMethodList, which stops at the superclass
boundary -- therefore, it wouldn't see the methods that we added to
RCTViewManager in an @implementation.