CircleCI deploy from tags rather than master
This updates our CircleCI workflow to run build/test on `v*` branches (e.g. `v1.1.0`) in addition to all branches, and then it moves the hold and deployment jobs to *only* run on `v*` branches.

This is loosely based on https://circleci.com/blog/publishing-npm-packages-using-circleci-2-0/

This should resolve the issue we're currently experiencing where the CI status badge is marked as failing even though we've fixed the issue.

Unfortunately, this stuff is difficult to test until we're ready to do an actual deployment, but I think I got it right.