Synchronize invoking the onReady callback (#122)
* Synchronize invoking the onReady callback

A customer experienced a rare crash that seems consistent with a race in accessing the resolve callback passed in from the javascript onReady API. If called at just the right time, the delegate could try to invoke teh callback at the same time as the code setting the callback, leading to it being nulled out just before one of them calls it.

* race condition safety (#123)

* Synchronize invoking the onReady callback

A customer experienced a rare crash that seems consistent with a race in accessing the resolve callback passed in from the javascript onReady API. If called at just the right time, the delegate could try to invoke teh callback at the same time as the code setting the callback, leading to it being nulled out just before one of them calls it.

* race condition safety

---------

Co-authored-by: Peter Fedak <peterfedak@fullstory.com>

---------

Co-authored-by: Ryan Wang <qwang1993@gmail.com>